some bug fixes updated new IDs in files
git-svn-id: http://svn.xp-dev.com/svn/Sweil_fs2/src@116 b49b0359-402e-0410-ad19-9fd90e479030

+++ www/admin/admin_cimgdel.php
@@ -4,11 +4,11 @@ echo '<script type="text/javascript">
       {
         var Check = confirm("Soll das Bild wirklich gelöscht werden?");
         if (Check == true)
-          location.href = "?mid='.$_GET[mid].'&go=cimgdel&unlink=" + file + "'.$session_url.'";
+          location.href = "?go=cimg_admin&unlink=" + file + "'.$session_url.'";
       }
       </script>';
 
-if (isset($_GET['unlink']))
+if (isset($_GET['unlink']) && $_SESSION['cimg_admin'] == 1 )
 {
   unlink("../images/content/".$_GET['unlink']);
   systext("Das Bild \"".$_GET['unlink']."\" wurde gelöscht!");
